+++ src/event/input.rs
@@ -1,18 +1,20 @@
-use crossterm::event::poll;
 use crossterm::event::Event;
+use crossterm::event::EventStream;
 use crossterm::event::KeyCode;
 use crossterm::event::KeyEvent;
 use crossterm::event::KeyModifiers;
-use crossterm::event::{
-    self,
-};
-use std::thread;
+use futures::FutureExt;
+use futures::StreamExt;
+use log::info;
 use std::time::Duration;
 use tokio::net::TcpStream;
+use tokio::select;
 use tokio::sync::mpsc::Sender;
+use tokio::task;
 
 use crate::app::SelectedView;
 use crate::dbgp::client::ContinuationStatus;
+use crate::notification::Notification;
 use crate::view::session::SessionViewMode;
 use crate::view::Scroll;
 
@@ -55,32 +57,55 @@ pub enum AppEvent {
     EvalExecute,
     EvalRefresh,
     EvalStart,
+    Listening,
+    NextChannel,
+    FocusChannel(String),
+    ChannelLog(String, String),
+    RestartProcess,
+    NotifyError(String),
+    Notify(Notification),
 }
 
 pub type EventSender = Sender<AppEvent>;
 
 pub fn start(event_sender: EventSender) {
-    thread::spawn(move || {
-        event_sender.blocking_send(AppEvent::Startup).unwrap();
+    let sender = event_sender.clone();
+    task::spawn(async move {
+        sender.send(AppEvent::Startup).await.unwrap();
+        let mut reader = EventStream::new();
+        let mut tick_interval = tokio::time::interval(Duration::from_millis(10));
         loop {
-            if poll(Duration::from_millis(1000)).unwrap() {
-                // handle global keys
-                if let Event::Key(key) = event::read().unwrap() {
-                    let action: Option<AppEvent> = match key.modifiers {
-                        KeyModifiers::CONTROL => match key.code {
-                            KeyCode::Char('c') => Some(AppEvent::Quit),
-                            _ => None,
-                        },
-                        _ => None,
-                    };
+            let tick = tick_interval.tick().fuse();
+            let event = reader.next().fuse();
 
-                    match action {
-                        Some(a) => event_sender.blocking_send(a).unwrap(),
-                        None => event_sender.blocking_send(AppEvent::Input(key)).unwrap(),
+            select! {
+                _ = tick => {
+                    let _ = sender.send(AppEvent::Tick).await;
+                },
+                maybe_event = event => {
+                    match maybe_event {
+                        Some(Ok(Event::Key(key))) => {
+                            info!("{:?}", key);
+                            let action: Option<AppEvent> = match key.modifiers {
+                                KeyModifiers::CONTROL => match key.code {
+                                    KeyCode::Char('c') => Some(AppEvent::Quit),
+                                    _ => None,
+                                },
+                                _ => None,
+                            };
+
+                            match action {
+                                Some(a) => event_sender.send(a).await.unwrap(),
+                                None => event_sender.send(AppEvent::Input(key)).await.unwrap(),
+                            }
+                        }
+                        Some(Err(e)) => {
+                            event_sender.send(AppEvent::NotifyError(e.to_string())).await.unwrap()
+                        },
+                        _ => {},
                     }
                 }
             }
-            event_sender.blocking_send(AppEvent::Tick).unwrap();
         }
     });
 }